+++ src/dp_matrix.h
@@ -25,45 +25,41 @@ struct dp_matrix
 
 struct dp_matrix*  dp_matrix_create(struct dp_state_table const* states);
 void               dp_matrix_destroy(struct dp_matrix const* matrix);
-static inline void dp_matrix_get_prev_step(struct dp_matrix const* dp_matrix, uint_fast16_t row,
-                                           uint_fast16_t state, uint_fast16_t* step_state,
-                                           uint_fast8_t* step_seq_len);
-static inline void dp_matrix_set_prev_step(struct dp_matrix const* dp_matrix, uint_fast16_t row,
-                                           uint_fast16_t state, uint_fast16_t step_state,
-                                           uint_fast8_t step_seq_len);
-static inline imm_float dp_matrix_get_score(struct dp_matrix const* dp_matrix, uint_fast16_t row,
-                                            uint_fast16_t state, uint_fast8_t seq_len);
-static inline void      dp_matrix_set_score(struct dp_matrix const* dp_matrix, uint_fast16_t row,
-                                            uint_fast16_t state, uint_fast8_t seq_len, imm_float score);
+static inline void dp_matrix_get_prev_step(struct dp_matrix const* dp_matrix, uint_fast32_t row, uint_fast16_t state,
+                                           uint_fast16_t* step_state, uint_fast8_t* step_seq_len);
+static inline void dp_matrix_set_prev_step(struct dp_matrix const* dp_matrix, uint_fast32_t row, uint_fast16_t state,
+                                           uint_fast16_t step_state, uint_fast8_t step_seq_len);
+static inline imm_float dp_matrix_get_score(struct dp_matrix const* dp_matrix, uint_fast32_t row, uint_fast16_t state,
+                                            uint_fast8_t seq_len);
+static inline void      dp_matrix_set_score(struct dp_matrix const* dp_matrix, uint_fast32_t row, uint_fast16_t state,
+                                            uint_fast8_t seq_len, imm_float score);
 void                    dp_matrix_setup(struct dp_matrix* matrix, struct eseq const* eseq);
 
-static inline void dp_matrix_get_prev_step(struct dp_matrix const* dp_matrix, uint_fast16_t row,
-                                           uint_fast16_t state, uint_fast16_t* step_state,
-                                           uint_fast8_t* step_seq_len)
+static inline void dp_matrix_get_prev_step(struct dp_matrix const* dp_matrix, uint_fast32_t row, uint_fast16_t state,
+                                           uint_fast16_t* step_state, uint_fast8_t* step_seq_len)
 {
     *step_state = matrixu16_get(dp_matrix->prev_state, row, state);
     *step_seq_len = matrixu8_get(dp_matrix->prev_seq_len, row, state);
 }
 
-static inline void dp_matrix_set_prev_step(struct dp_matrix const* dp_matrix, uint_fast16_t row,
-                                           uint_fast16_t state, uint_fast16_t step_state,
-                                           uint_fast8_t step_seq_len)
+static inline void dp_matrix_set_prev_step(struct dp_matrix const* dp_matrix, uint_fast32_t row, uint_fast16_t state,
+                                           uint_fast16_t step_state, uint_fast8_t step_seq_len)
 {
     matrixu16_set(dp_matrix->prev_state, row, state, (uint16_t)step_state);
     matrixu8_set(dp_matrix->prev_seq_len, row, state, (uint8_t)step_seq_len);
 }
 
-static inline imm_float dp_matrix_get_score(struct dp_matrix const* dp_matrix, uint_fast16_t row,
-                                            uint_fast16_t state, uint_fast8_t seq_len)
+static inline imm_float dp_matrix_get_score(struct dp_matrix const* dp_matrix, uint_fast32_t row, uint_fast16_t state,
+                                            uint_fast8_t seq_len)
 {
-    uint_fast16_t col = (uint_fast16_t)(dp_matrix->state_col[state] + (int_fast16_t)seq_len);
+    uint_fast32_t col = (uint_fast32_t)(dp_matrix->state_col[state] + (int_fast32_t)seq_len);
     return matrixf_get(dp_matrix->score, row, col);
 }
 
-static inline void dp_matrix_set_score(struct dp_matrix const* dp_matrix, uint_fast16_t row,
-                                       uint_fast16_t state, uint_fast8_t seq_len, imm_float score)
+static inline void dp_matrix_set_score(struct dp_matrix const* dp_matrix, uint_fast32_t row, uint_fast16_t state,
+                                       uint_fast8_t seq_len, imm_float score)
 {
-    uint_fast16_t col = (uint_fast16_t)(dp_matrix->state_col[state] + (int_fast16_t)seq_len);
+    uint_fast32_t col = (uint_fast32_t)(dp_matrix->state_col[state] + (int_fast32_t)seq_len);
     matrixf_set(dp_matrix->score, row, col, score);
 }
